EquivalenceClassesConstructor

EquivalenceClassesConstructor.EquivalenceClasses

EquivalenceClassesConstructor.build_adj_matrix

EquivalenceClassesConstructor.check_for_equivalence_relation

EquivalenceClassesConstructor.make_symmetric

EquivalenceClassesConstructor.EquivalenceClasses — Method

EquivalenceClasses(pred, indl)

Constructs an EquivalenceClasses struct from a predicate pred with indices in the list indl

Examples

julia> EquivalenceClasses(Predicate((x,y)->all(x .== -y)),
                          [(i,j) for i in -2:2 for j in 4:7])

EquivalenceClassesConstructor.build_adj_matrix — Method

build_adj_matrix(pred, indl)

Constructs adjacency matrix with vertices from indl and edges from pred(i,j) where i,j ∈ indl

EquivalenceClassesConstructor.check_for_equivalence_relation — Method

check_for_equivalence_relation(f::Predicate, dom)

Given a Predicate f over a domain dom, return true if f is a equivalence relation, false otherwise. One can create a symmetric equivalency relation with make_symmetric(f)

Examples

true
julia> check_for_equivalence_relation(Predicate((x,y) -> x == y+2), -1:10)
false

EquivalenceClassesConstructor.make_symmetric — Method

make_symmetric(f::Predicate)

Create symmetric predicate from arbitrary predicateg