EquivalenceClassesConstructor

EquivalenceClassesConstructor.EquivalenceClasses

EquivalenceClassesConstructor.build_adj_matrix

EquivalenceClassesConstructor.check_for_equivalence_relation

EquivalenceClassesConstructor.find_classes

EquivalenceClassesConstructor.find_classes

EquivalenceClassesConstructor.EquivalenceClasses — Method

EquivalenceClasses(pred, indl)

Constructs an EquivalenceClasses struct from a predicate pred with indices in the list indl. pred needs to be a equivalence relation: reflexivity, symmetry and transitivity will be enforced by the algorithm creating the adjecency matrix.

Examples

julia> EquivalenceClasses(Predicate((x,y)->all(x .== -y)),
                          [(i,j) for i in -2:2 for j in 4:7])

EquivalenceClassesConstructor.build_adj_matrix — Method

build_adj_matrix(pred, indl)

Constructs adjacency matrix with vertices from indl and edges from pred(i,j) where i,j ∈ indl

EquivalenceClassesConstructor.check_for_equivalence_relation — Method

check_for_equivalence_relation(f::Predicate, dom)

Given a Predicate f over a domain dom, return true if f is a equivalence relation, false otherwise. One can create a symmetric equivalency relation with make_symmetric(f) TODO: check transitivity.

Examples

true
julia> check_for_equivalence_relation(Predicate((x,y) -> x == y+2), -1:10)
false

EquivalenceClassesConstructor.find_classes — Method

find_classes(adj::BitArray{2})

returns an array of length size(adj,1) with each entry with index i being a unique identifier for the equivalency class of the node i.

EquivalenceClassesConstructor.find_classes — Method

find_classes(pred::Predicate, indl)

returns an array of length length(indl) with each entry with index i being a unique identifier for the equivalency class of the node i.