EquivalenceClassesConstructor

This code aims to provide a collection of methods that are able to exploit symmetries in arbitrary indexable objects over a given index set vl. For any indices x, y these symmetries must be known in the form of predicates, i.e. p(x,y) &#8712; {true, false} or in form of mappings m(x) = y. The latter greatly increases performance but may lead to unexpected behavior with non monotonic mappings. The reason being that m(m(x)) may lie in vl but m(x)) may not, forcing the algorithm to terminate.

Index

EquivalenceClassesConstructor.EquivalenceClasses

EquivalenceClassesConstructor.EquivalenceClasses

EquivalenceClassesConstructor.build_adj_matrix

EquivalenceClassesConstructor.check_for_equivalence_relation

EquivalenceClassesConstructor.find_classes

EquivalenceClassesConstructor.find_classes

EquivalenceClassesConstructor.find_classes

EquivalenceClassesConstructor.EquivalenceClasses — Method

EquivalenceClasses(m::Mapping, vl::AbstractArray)

Constructs an EquivalenceClasses struct from a mapping m over the indices in the list vl. sorted can be set to true to force all lists of mappings between classes to be sorted

Examples

julia> EquivalenceClasses(Predicate((x,y)->all(x .== -y)),
                          [(i,j) for i in -2:2 for j in 4:7])

EquivalenceClassesConstructor.EquivalenceClasses — Method

EquivalenceClasses(pred::Predicate, vl::AbstractArray)

Constructs an EquivalenceClasses struct from a predicate pred over n indices in the list vl. This inputs requires n^2 operations in order to create the adjecency matrix. For large inputs a mapping is therefore preferabel. sorted can be set to true to force all lists of mappings between classes to be sorted

Examples

julia> EquivalenceClasses(Predicate((x,y)->all(x .== -y)),
                          [(i,j) for i in -2:2 for j in 4:7])

EquivalenceClassesConstructor.build_adj_matrix — Method

build_adj_matrix(pred, vl)

Constructs adjacency matrix with vertices from vl and edges from pred(i,j) where i,j ∈ vl

EquivalenceClassesConstructor.check_for_equivalence_relation — Method

check_for_equivalence_relation(f::Predicate, dom)

Given a Predicate f over a domain dom, return true if f is a equivalence relation, false otherwise. One can create a symmetric equivalency relation with make_symmetric(f) TODO: check transitivity.

Examples

julia> check_for_equivalence_relation(Predicate((x,y) -> x == -y), -5:10)
true
julia> check_for_equivalence_relation(Predicate((x,y) -> x == y+2), -1:10)
false

EquivalenceClassesConstructor.find_classes — Method

find_classes(adj::BitArray{2})

returns an array of length size(adj,1) with each entry with index i being a unique identifier for the equivalency class of the node i. Uses adjacency matrix as input. I.e. graph needs to be constructed before.

Examples

julia> find_classes(convert(BitArray, [0 1 0; 1 0 0; 0 0 0]))
[1, 1, 2]

EquivalenceClassesConstructor.find_classes — Method

find_classes(pred::Predicate, vl; isSymmetric=false)

returns an array of length length(vl) with each entry with index i being a unique identifier for the equivalency class of the node i. See find_classes(m::Mapping, vl) for more information.

EquivalenceClassesConstructor.find_classes — Method

find_classes(m::Mapping, vl)

returns an array of length size(adj,1) with each entry with index i being a unique identifier for the equivalency class of the node i. Uses a mapping function m(x) = y and an vertex list vl to construct graph i n place. m has to be symmetric (i.e. m(x) = y => m(y) = x)! In case the vertex list is closed under the mapping (i.e. there exist no x in vl such that m(x) is not in vl), closed can be set to false in order to improve performace. The boolean sorted can be set to true, if the resulting expand map should be sorted (one should make sure, that the elements of vl are comparable). For performace reasons this function can also be called with a normal Julia function instead of Mapping

Examples

julia> find_classes(Mapping(x-> [-x]), -2:4)
"IndexMapping{Int64}[Full: -2:4, ExpandMap: Dict(0 => [0],4 => [4],-2 => [-2, 2],-1 => [-1, 1],3 => [3])]"